Add pluggable web server lifecycle
