+++ docs/reference/faust.contrib.asgi.rst
@@ -1,0 +1,9 @@
+=====================================================
+ ``faust.contrib.asgi``
+=====================================================
+
+.. currentmodule:: faust.contrib.asgi
+
+.. automodule:: faust.contrib.asgi
+    :members:
+    :undoc-members:
