+++ docs/reference/faust.contrib.fastapi.rst
@@ -2,6 +2,9 @@
  ``faust.contrib.fastapi``
 =====================================================
 
+Compatibility alias for :mod:`faust.contrib.asgi`. New code should import the
+framework-neutral module directly.
+
 .. contents::
     :local:
 .. currentmodule:: faust.contrib.fastapi
