+++ docs/reference/faust.sensors.metrics.rst
@@ -1,0 +1,11 @@
+=====================================================
+ ``faust.sensors.metrics``
+=====================================================
+
+.. contents::
+    :local:
+.. currentmodule:: faust.sensors.metrics
+
+.. automodule:: faust.sensors.metrics
+    :members:
+    :undoc-members:
