+++ faust/sensors/__init__.py
@@ -1,10 +1,12 @@
 """Sensors."""
 
 from .base import Sensor, SensorDelegate
+from .metrics import performance_metrics
 from .monitor import Monitor, TableState
 
 __all__ = [
     "Monitor",
+    "performance_metrics",
     "Sensor",
     "SensorDelegate",
     "TableState",
