+++ requirements/extras/asgi.txt
@@ -1,0 +1,1 @@
+uvicorn>=0.27.0
