+++ requirements/extras/fastapi.txt
@@ -1,2 +1,2 @@
 fastapi>=0.100.0
-uvicorn>=0.27.0
+-r asgi.txt
