+++ requirements/test.txt
@@ -45,7 +45,7 @@ intervaltree
 opentelemetry-api>=1.20.0
 opentelemetry-sdk>=1.20.0
 # uvicorn's only dependencies are click (already required by faust) and h11,
-# both pure Python -- cheap enough to cover faust.contrib.fastapi's server
+# both pure Python -- cheap enough to cover faust.contrib.asgi's server
 # construction on every matrix leg.  fastapi itself is not added here: it
 # pulls in pydantic-core, which has no wheel for every leg.
 uvicorn>=0.27.0
